requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-filter==23.1
django-rest-swagger==2.2.0
django-shell-plus==1.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-nested-routers==0.93.4
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.11
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.19.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
yarg==0.1.9
